+++ src/_gettsim/ids.py
@@ -67,21 +67,21 @@ def fg_id(
 
     # Assign the same fg_id to everybody who has an Einstandspartner,
     # otherwise create a new one from p_id
-    fg_id = np.where(
+    out = np.where(
         arbeitslosengeld_2__p_id_einstandspartner < 0,
         p_id + p_id * n,
         np.maximum(p_id, arbeitslosengeld_2__p_id_einstandspartner)
         + np.minimum(p_id, arbeitslosengeld_2__p_id_einstandspartner) * n,
     )
 
-    fg_id = _assign_parents_fg_id(
-        fg_id, p_id, p_id_elternteil_1_loc, hh_id, alter, children, n
+    out = _assign_parents_fg_id(
+        out, p_id, p_id_elternteil_1_loc, hh_id, alter, children, n
     )
-    fg_id = _assign_parents_fg_id(
-        fg_id, p_id, p_id_elternteil_2_loc, hh_id, alter, children, n
+    out = _assign_parents_fg_id(
+        out, p_id, p_id_elternteil_2_loc, hh_id, alter, children, n
     )
 
-    return fg_id
+    return out
 
 
 def _assign_parents_fg_id(
@@ -135,13 +135,12 @@ def bg_id(
     # https://github.com/iza-institute-of-labor-economics/gettsim/issues/668
     offset = np.max(fg_id) + 1
     # Create new id for everyone who is not part of the Bedarfsgemeinschaft
-    bg_id = np.where(
+
+    return np.where(
         (arbeitslosengeld_2__eigenbedarf_gedeckt) * (alter < 25),
         offset + p_id,
         fg_id,
     )
-
-    return bg_id
 
 
 @group_creation_function()
@@ -159,12 +158,11 @@ def eg_id(
         p_id,
         arbeitslosengeld_2__p_id_einstandspartner,
     )
-    result = (
+
+    return (
         np.maximum(p_id, p_id_einstandspartner__or_own_p_id)
         + np.minimum(p_id, p_id_einstandspartner__or_own_p_id) * n
     )
-
-    return result
 
 
 @group_creation_function()
@@ -179,13 +177,13 @@ def wthh_id(
     priority check compared to Bürgergeld yields the same result ∈ {True, False}.
     """
     offset = np.max(hh_id) + 1
-    wthh_id = np.where(
+
+    return np.where(
         vorrangprüfungen__wohngeld_vorrang_vor_arbeitslosengeld_2_bg
         | vorrangprüfungen__wohngeld_und_kinderzuschlag_vorrang_vor_arbeitslosengeld_2_bg,
         hh_id + offset,
         hh_id,
     )
-    return wthh_id
 
 
 @group_creation_function()
@@ -207,9 +205,7 @@ def sn_id(
         p_id,
     )
 
-    result = (
+    return (
         np.maximum(p_id, p_id_ehepartner_or_own_p_id)
         + np.minimum(p_id, p_id_ehepartner_or_own_p_id) * n
     )
-
-    return result
